Update import statments in notebooks 2
